server/pyproject.toml
[project]
name = "parking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "geopy>=2.4.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "python-decouple>=3.8",
]
